Edit requestMatcher parameter of FilterChain in SecurityConfig
1. edit parameter of the authorizeHttpRequests process

At the beginning of the development, I've set whole kinds of mapping can access without any check.

But that is serious problem to be hacked, so I separate two ways. which are one for some pages requeired login process and the other one for anyone can access without login process

2. edit login process in SecurityFilterChain

When use login process through Spring Security, the parameter of loginPage must be "/login". And I allow user to access in the myPage when user succeeded in the login, but if there are some error by system or user during login process the system will bring user to be back login form page, which is mapped by "/member".

3. add logout process in SecurityFilterChain

When the logout process has done, the system will carry user to the Item List page.

4. Also, remove ex-version of login page and switch to new version of login page