Show attached image into the Item List board
After image upload function in item add form has successfully made, it's time to put uploaded image into Item List board.

1. To save uploaded image url, make input tag into form tag.
The input tag must set that the value of type is hidden and the value of id and name is "imageUrl".

2. In imageS3upload javascript code,  there is the result in the transfer part to the add item form page.

uploadImage plays for the image url store in S3 bucket. And imageUrl plays for the image url store in DB for show up to Item List board.