Exception Handling - DELETE process
If user try to delete certain item from the Item List board, the server will check the log-in status before the deleting item process.

And then, if user has not logged into the site, the server will announce several message to user. The message demands user to do log-in progress.

But the displayName in user's member log is not same as the uploader name or admin, the server will deny to delete item.