Exception Handling - DELETE process 2
Before remove item from the Item List board, the server will check that user has logged in or not.

If user has not logged in yet, the server will show some message to announce that user must log in if user wants to delete item. And then the server brings user to login form.

When user does log in, the display name of user will show up in the navigation bar.

And user try to erase specific item from Item List board, the server will check user 's display name is same as uploader's name or user is admin. If user's display name has fit to uploader's name or admin, the item will successfully remove.