Move the function to go to detail view page
The function that user can go to item detail view page was made in "선택 기능" of item board.

But mostly the way to access item detail view page is to move through title or image columns.

So, relocate the way to go to item detail view page into the title and image column of the item board. Also, notice board set same function in the title column.