Indicate user's choice in navigation bar
When user has selected menu in navigation bar, it has not pointed up the page clearly what user has accessed before check the url.

Therefore, create a javascript to add active state in the item of menu, and make a css to change the font color and font-weight when user select.

Also, the javascript and css of the navigation bar is common factor in the whole pages. So, put them into the common fragments to manage well.