Check the login state of user
When user use board functions like item add or delete, the system must check user state like log in or out.

Because the function to create, edit, delete item does not allow for anonymous user. This  is related in security of the site.

Therefore, the check user state on session is very important.

Also, to aboid code duplication, the process of Authentication check has separated as an individual step.

In SecurityConfig, the filter has classified unauthorize user or access, the exceptionHandling will bring to AuthenticationInterceptor which implements HandlerInterceptor.