Fix Pagination bug
The pagination process has not managed the change of board's total page when item has added or deleted.

Make some methods to check increase or decrease total page of board in CacheService.

Put increaseItemCount method in add board item process and decreaseItemCount method in delete board item process.

Now, the pagination controls in board recognize the change of total page and apply it in real time.