Add the cancel button for selected image
When user add new item in Item board, user is writing name, price and the image of item.

This function is for the case if user uploaded wrong image in item add form. The point of time when user has chosen and selected an image file, the image file has already uploaded in S3 bucket.

Thus, this button works to erase uploaded wrong image file from S3 bucket.