The main contents of my page part
The HTML pages related in my page part are now using the layout of my page part.

And now look at the progress how user accesses into my page part of their own account.

First, after user logs in the site and the site shows the information of user's account on the right side of navigation bar.

Second, when user clicks their account's name, the recheck procedure begins to work. It's to check the user is really the owner of the account.

Thus, the re-login form shows their account's id and demands user to repeat to type the password of their own account. If the user is right, the real my page opens. If not, the access denies and get back to the main page before user has logged in.

Third, the main page of my page has side menu, and the side menu is only for my page.
The side menu locates in the left side from the main contents of my page.

Oh, myPage_debug.html is just for developer when check the user information has shown and tranferred well when developing my page. Yeah, this file is only for debugging in browser.