Set back-end process of my page part 1
1. SecurityConfig
put some details in url value for authorized Http request has happened from my page part.

This method demands password what user has already typed in the input box. Transfer the value of username and password to MemberService which handles the business logic of each method in MemberController.

If the check is true, the method does redirect to main page of my-page part which calls showMyPage method in same Controller class.

If not, the user gets back to re-login form page again. Which is called checkMyPage method in same Controller class.