Set back-end process of my-page part 2
2.  Rebuild the process of my page access
remove the prototype of my page accessing, and split up the prototype progress.

a. showMyPage
This method handles main control to access my-page. It checks the value of "isVerified". This condition prevents that stranger tries to get into other's my-page page.

If the value of "isVerified" is true, the method is aware of the owener's request. So, the method brings the user to their own my-page page.

Also, the value of "isVerified" removes because this value is temporary condition.

b. checkMyPage
This method shows the re-Login form to user. Also, set the value of showSidebar to false because side bar sets only in main contents of my page part.

c. verifyUser
This method works the process of recheck that the user is really the owner of the account.