Refactory SecurityConfig class 1
As the SecurityConfig code is getting longer and become more complicated, it's hard to find that a certain method or variable locates in the class.

So, I decide to make the class be more simple. It's the first step of refactory the class. I separate the method in more detail.